@@ -3,23 +3,23 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
-  oktoast: ^2.3.0
-  image: ^2.1.4
+  cupertino_icons: ^1.0.3
+  oktoast: ^3.0.0
+  image: ^3.0.2
   esc_pos_bluetooth:
     path: ../../
-  esc_pos_utils: ^1.0.0
+  esc_pos_utils: ^1.1.0
   # esc_pos_utils:
   # path: ../../../esc_pos_utils
-  charset_converter: ^1.0.3
-  intl: ^0.16.1
-  qr_flutter: ^3.2.0
-  path_provider: ^1.6.5
+  charset_converter: ^2.0.0
+  intl: ^0.17.0
+  qr_flutter: ^4.0.0
+  path_provider: ^2.0.2
 
 dev_dependencies:
   flutter_test:
