@@ -1,19 +1,19 @@
 name: esc_pos_bluetooth
 description: The library allows to print receipts using an ESC/POS thermal Bluetooth printer.
-version: 0.2.8
+version: 0.4.1
 homepage: https://github.com/andrey-ushakov/esc_pos_bluetooth
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  rxdart: ^0.23.1
-  esc_pos_utils: ^1.0.0
+  rxdart: ^0.26.0
+  esc_pos_utils: ^1.1.0
   # esc_pos_utils:
   # path: ../esc_pos_utils
-  flutter_bluetooth_basic: ^0.1.5
+  flutter_bluetooth_basic: ^0.1.7
 
 dev_dependencies:
   flutter_test:
