@@ -28,7 +28,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
   final String title;
 
   @override
@@ -62,19 +62,18 @@ class _MyHomePageState extends State<MyHomePage> {
     printerManager.stopScan();
   }
 
-  Future<List<int>> demoReceipt(PaperSize paper) async {
-    final Generator generator =
-        Generator(paper, await CapabilityProfile.load());
+  Future<List<int>> demoReceipt(
+      PaperSize paper, CapabilityProfile profile) async {
+    final Generator ticket = Generator(paper, profile);
+    List<int> bytes = [];
 
     // Print image
-    final ByteData data = await rootBundle.load('assets/rabbit_black.jpg');
-    final Uint8List bytes = data.buffer.asUint8List();
-    final Image image = decodeImage(bytes);
-    // ticket.image(image);
+    // final ByteData data = await rootBundle.load('assets/rabbit_black.jpg');
+    // final Uint8List imageBytes = data.buffer.asUint8List();
+    // final Image? image = decodeImage(imageBytes);
+    // bytes += ticket.image(image);
 
-    List<int> receiptBytes = [];
-
-    receiptBytes += generator.text('GROCERYLY',
+    bytes += ticket.text('GROCERYLY',
         styles: PosStyles(
           align: PosAlign.center,
           height: PosTextSize.size2,
@@ -82,17 +81,17 @@ class _MyHomePageState extends State<MyHomePage> {
         ),
         linesAfter: 1);
 
-    receiptBytes += generator.text('889  Watson Lane',
+    bytes += ticket.text('889  Watson Lane',
         styles: PosStyles(align: PosAlign.center));
-    receiptBytes += generator.text('New Braunfels, TX',
+    bytes += ticket.text('New Braunfels, TX',
         styles: PosStyles(align: PosAlign.center));
-    receiptBytes += generator.text('Tel: 830-221-1234',
+    bytes += ticket.text('Tel: 830-221-1234',
         styles: PosStyles(align: PosAlign.center));
-    receiptBytes += generator.text('Web: www.example.com',
+    bytes += ticket.text('Web: www.example.com',
         styles: PosStyles(align: PosAlign.center), linesAfter: 1);
 
-    receiptBytes += generator.hr();
-    receiptBytes += generator.row([
+    bytes += ticket.hr();
+    bytes += ticket.row([
       PosColumn(text: 'Qty', width: 1),
       PosColumn(text: 'Item', width: 7),
       PosColumn(
@@ -101,7 +100,7 @@ class _MyHomePageState extends State<MyHomePage> {
           text: 'Total', width: 2, styles: PosStyles(align: PosAlign.right)),
     ]);
 
-    receiptBytes += generator.row([
+    bytes += ticket.row([
       PosColumn(text: '2', width: 1),
       PosColumn(text: 'ONION RINGS', width: 7),
       PosColumn(
@@ -109,7 +108,7 @@ class _MyHomePageState extends State<MyHomePage> {
       PosColumn(
           text: '1.98', width: 2, styles: PosStyles(align: PosAlign.right)),
     ]);
-    receiptBytes += generator.row([
+    bytes += ticket.row([
       PosColumn(text: '1', width: 1),
       PosColumn(text: 'PIZZA', width: 7),
       PosColumn(
@@ -117,7 +116,7 @@ class _MyHomePageState extends State<MyHomePage> {
       PosColumn(
           text: '3.45', width: 2, styles: PosStyles(align: PosAlign.right)),
     ]);
-    receiptBytes += generator.row([
+    bytes += ticket.row([
       PosColumn(text: '1', width: 1),
       PosColumn(text: 'SPRING ROLLS', width: 7),
       PosColumn(
@@ -125,7 +124,7 @@ class _MyHomePageState extends State<MyHomePage> {
       PosColumn(
           text: '2.99', width: 2, styles: PosStyles(align: PosAlign.right)),
     ]);
-    receiptBytes += generator.row([
+    bytes += ticket.row([
       PosColumn(text: '3', width: 1),
       PosColumn(text: 'CRUNCHY STICKS', width: 7),
       PosColumn(
@@ -133,9 +132,9 @@ class _MyHomePageState extends State<MyHomePage> {
       PosColumn(
           text: '2.55', width: 2, styles: PosStyles(align: PosAlign.right)),
     ]);
-    receiptBytes += generator.hr();
+    bytes += ticket.hr();
 
-    receiptBytes += generator.row([
+    bytes += ticket.row([
       PosColumn(
           text: 'TOTAL',
           width: 6,
@@ -153,9 +152,9 @@ class _MyHomePageState extends State<MyHomePage> {
           )),
     ]);
 
-    receiptBytes += generator.hr(ch: '=', linesAfter: 1);
+    bytes += ticket.hr(ch: '=', linesAfter: 1);
 
-    receiptBytes += generator.row([
+    bytes += ticket.row([
       PosColumn(
           text: 'Cash',
           width: 7,
@@ -165,7 +164,7 @@ class _MyHomePageState extends State<MyHomePage> {
           width: 5,
           styles: PosStyles(align: PosAlign.right, width: PosTextSize.size2)),
     ]);
-    receiptBytes += generator.row([
+    bytes += ticket.row([
       PosColumn(
           text: 'Change',
           width: 7,
@@ -176,14 +175,14 @@ class _MyHomePageState extends State<MyHomePage> {
           styles: PosStyles(align: PosAlign.right, width: PosTextSize.size2)),
     ]);
 
-    receiptBytes += generator.feed(2);
-    receiptBytes += generator.text('Thank you!',
+    bytes += ticket.feed(2);
+    bytes += ticket.text('Thank you!',
         styles: PosStyles(align: PosAlign.center, bold: true));
 
     final now = DateTime.now();
     final formatter = DateFormat('MM/dd/yyyy H:m');
     final String timestamp = formatter.format(now);
-    receiptBytes += generator.text(timestamp,
+    bytes += ticket.text(timestamp,
         styles: PosStyles(align: PosAlign.center), linesAfter: 2);
 
     // Print QR Code from image
@@ -201,41 +200,43 @@ class _MyHomePageState extends State<MyHomePage> {
     //   final imgFile = await qrFile.writeAsBytes(uiImg.buffer.asUint8List());
     //   final img = decodeImage(imgFile.readAsBytesSync());
 
-    //   ticket.image(img);
+    //   bytes += ticket.image(img);
     // } catch (e) {
     //   print(e);
     // }
 
     // Print QR Code using native function
-    // ticket.qrcode('example.com');
+    // bytes += ticket.qrcode('example.com');
 
-    receiptBytes += generator.feed(2);
-    receiptBytes += generator.cut();
-    return receiptBytes;
+    ticket.feed(2);
+    ticket.cut();
+    return bytes;
   }
 
-  Future<List<int>> testTicket(PaperSize paper) async {
-    final Generator ticket = Generator(paper, await CapabilityProfile.load());
-    List<int> receiptBytes = [];
+  Future<List<int>> testTicket(
+      PaperSize paper, CapabilityProfile profile) async {
+    final Generator generator = Generator(paper, profile);
+    List<int> bytes = [];
 
-    receiptBytes += ticket.text(
+    bytes += generator.text(
         'Regular: aA bB cC dD eE fF gG hH iI jJ kK lL mM nN oO pP qQ rR sS tT uU vV wW xX yY zZ');
-    receiptBytes += ticket.text('Special 1: àÀ èÈ éÉ ûÛ üÜ çÇ ôÔ');
-    receiptBytes += ticket.text('Special 2: blåbærgrød');
+    // bytes += generator.text('Special 1: àÀ èÈ éÉ ûÛ üÜ çÇ ôÔ',
+    //     styles: PosStyles(codeTable: PosCodeTable.westEur));
+    // bytes += generator.text('Special 2: blåbærgrød',
+    //     styles: PosStyles(codeTable: PosCodeTable.westEur));
 
-    receiptBytes += ticket.text('Bold text', styles: PosStyles(bold: true));
-    receiptBytes +=
-        ticket.text('Reverse text', styles: PosStyles(reverse: true));
-    receiptBytes += ticket.text('Underlined text',
+    bytes += generator.text('Bold text', styles: PosStyles(bold: true));
+    bytes += generator.text('Reverse text', styles: PosStyles(reverse: true));
+    bytes += generator.text('Underlined text',
         styles: PosStyles(underline: true), linesAfter: 1);
-    receiptBytes +=
-        ticket.text('Align left', styles: PosStyles(align: PosAlign.left));
-    receiptBytes +=
-        ticket.text('Align center', styles: PosStyles(align: PosAlign.center));
-    receiptBytes += ticket.text('Align right',
+    bytes +=
+        generator.text('Align left', styles: PosStyles(align: PosAlign.left));
+    bytes += generator.text('Align center',
+        styles: PosStyles(align: PosAlign.center));
+    bytes += generator.text('Align right',
         styles: PosStyles(align: PosAlign.right), linesAfter: 1);
 
-    receiptBytes += ticket.row([
+    bytes += generator.row([
       PosColumn(
         text: 'col3',
         width: 3,
@@ -253,7 +254,7 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
     ]);
 
-    receiptBytes += ticket.text('Text size 200%',
+    bytes += generator.text('Text size 200%',
         styles: PosStyles(
           height: PosTextSize.size2,
           width: PosTextSize.size2,
@@ -261,28 +262,28 @@ class _MyHomePageState extends State<MyHomePage> {
 
     // Print image
     final ByteData data = await rootBundle.load('assets/logo.png');
-    final Uint8List bytes = data.buffer.asUint8List();
-    final Image image = decodeImage(bytes);
-    receiptBytes += ticket.image(image);
+    final Uint8List buf = data.buffer.asUint8List();
+    final Image image = decodeImage(buf)!;
+    bytes += generator.image(image);
     // Print image using alternative commands
-    // receiptBytes += ticket.imageRaster(image);
-    // receiptBytes += ticket.imageRaster(image, imageFn: PosImageFn.graphics);
+    // bytes += generator.imageRaster(image);
+    // bytes += generator.imageRaster(image, imageFn: PosImageFn.graphics);
 
     // Print barcode
     final List<int> barData = [1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 4];
-    receiptBytes += ticket.barcode(Barcode.upcA(barData));
+    bytes += generator.barcode(Barcode.upcA(barData));
 
     // Print mixed (chinese + latin) text. Only for printers supporting Kanji mode
-    // receiptBytes += ticket.text(
+    // bytes += generator.text(
     //   'hello ! 中文字 # world @ éphémère &',
     //   styles: PosStyles(codeTable: PosCodeTable.westEur),
     //   containsChinese: true,
     // );
 
-    receiptBytes += ticket.feed(2);
+    bytes += generator.feed(2);
 
-    receiptBytes += ticket.cut();
-    return receiptBytes;
+    bytes += generator.cut();
+    return bytes;
   }
 
   void _testPrint(PrinterBluetooth printer) async {
@@ -290,6 +291,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
     // TODO Don't forget to choose printer's paper
     const PaperSize paper = PaperSize.mm80;
+    final profile = await CapabilityProfile.load();
 
     // TEST PRINT
     // final PosPrintResult res =
@@ -297,7 +299,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
     // DEMO RECEIPT
     final PosPrintResult res =
-        await printerManager.printTicket(await demoReceipt(paper));
+        await printerManager.printTicket((await demoReceipt(paper, profile)));
 
     showToast(res.msg);
   }
@@ -329,7 +331,7 @@ class _MyHomePageState extends State<MyHomePage> {
                             mainAxisAlignment: MainAxisAlignment.center,
                             children: <Widget>[
                               Text(_devices[index].name ?? ''),
-                              Text(_devices[index].address),
+                              Text(_devices[index].address!),
                               Text(
                                 'Click to print a test receipt',
                                 style: TextStyle(color: Colors.grey[700]),
@@ -349,7 +351,7 @@ class _MyHomePageState extends State<MyHomePage> {
         stream: printerManager.isScanningStream,
         initialData: false,
         builder: (c, snapshot) {
-          if (snapshot.data) {
+          if (snapshot.data!) {
             return FloatingActionButton(
               child: Icon(Icons.stop),
               onPressed: _stopScanDevices,
