@@ -18,18 +18,19 @@ class PrinterBluetooth {
   PrinterBluetooth(this._device);
   final BluetoothDevice _device;
 
-  String get name => _device.name;
-  String get address => _device.address;
-  int get type => _device.type;
+  String? get name => _device.name;
+  String? get address => _device.address;
+  int? get type => _device.type;
 }
 
 /// Printer Bluetooth Manager
 class PrinterBluetoothManager {
   final BluetoothManager _bluetoothManager = BluetoothManager.instance;
   bool _isPrinting = false;
-  StreamSubscription _scanResultsSubscription;
-  StreamSubscription _isScanningSubscription;
-  PrinterBluetooth _selectedPrinter;
+  bool _isConnected = false;
+  StreamSubscription? _scanResultsSubscription;
+  StreamSubscription? _isScanningSubscription;
+  PrinterBluetooth? _selectedPrinter;
 
   final BehaviorSubject<bool> _isScanning = BehaviorSubject.seeded(false);
   Stream<bool> get isScanningStream => _isScanning.stream;
@@ -50,9 +51,9 @@ class PrinterBluetoothManager {
     _isScanningSubscription =
         _bluetoothManager.isScanning.listen((isScanningCurrent) async {
       // If isScanning value changed (scan just stopped)
-      if (_isScanning.value && !isScanningCurrent) {
-        _scanResultsSubscription?.cancel();
-        _isScanningSubscription?.cancel();
+      if (_isScanning.value! && !isScanningCurrent) {
+        _scanResultsSubscription!.cancel();
+        _isScanningSubscription!.cancel();
       }
       _isScanning.add(isScanningCurrent);
     });
@@ -69,7 +70,7 @@ class PrinterBluetoothManager {
   Future<PosPrintResult> writeBytes(List<int> bytes) async {
     if (_selectedPrinter == null) {
       return Future<PosPrintResult>.value(PosPrintResult.printerNotSelected);
-    } else if (_isScanning.value) {
+    } else if (_isScanning.value!) {
       return Future<PosPrintResult>.value(PosPrintResult.scanInProgress);
     } else if (_isPrinting) {
       return Future<PosPrintResult>.value(PosPrintResult.printInProgress);
@@ -80,6 +81,10 @@ class PrinterBluetoothManager {
     // We have to rescan before connecting, otherwise we can connect only once
     await _bluetoothManager.startScan(timeout: Duration(seconds: 1));
     await _bluetoothManager.stopScan();
+
+    // Connect
+    await _bluetoothManager.connect(_selectedPrinter!._device);
+
     // Subscribe to the events
     _bluetoothManager.state.listen((state) async {
       switch (state) {
@@ -94,23 +99,29 @@ class PrinterBluetoothManager {
       }
     });
 
-    // Connect
-    await _bluetoothManager
-        .connect(_selectedPrinter?._device ?? BluetoothDevice());
+    // Printing timeout
+    _runDelayed(timeout).then((dynamic v) async {
+      if (_isPrinting) {
+        _isPrinting = false;
+        completer.complete(PosPrintResult.timeout);
+      }
+    });
 
-    await _bluetoothManager.writeData(bytes);
-    sleep(Duration(seconds: 1));
-    try {
-      await _bluetoothManager.disconnect();
-    } catch (error) {
-      print('Failed to disconnect: $error');
-    }
-
-    _isPrinting = false;
-    return PosPrintResult.success;
+    return completer.future;
   }
 
-  Future<PosPrintResult> printTicket(List<int> bytes) async {
-    return writeBytes(bytes);
+  Future<PosPrintResult> printTicket(
+    List<int> bytes, {
+    int chunkSizeBytes = 20,
+    int queueSleepTimeMs = 20,
+  }) async {
+    if (bytes.isEmpty) {
+      return Future<PosPrintResult>.value(PosPrintResult.ticketEmpty);
+    }
+    return writeBytes(
+      bytes,
+      chunkSizeBytes: chunkSizeBytes,
+      queueSleepTimeMs: queueSleepTimeMs,
+    );
   }
 }
